new logic to create project allows more flexibility with db directory
