+++ setup/eatup_setup_server/Dockerfile
@@ -20,5 +20,14 @@ FROM alpine:3.16 as release
 WORKDIR /app
 # TODO copy scripts and needed resources form local
 # "/code/db/load_db.sql", "/installation/load_db.sql"
+
+# Installation files
+COPY ../../db /code # TODO this is not right
+
+# Frontend code
+COPY ../eatup_setup_front/dist /www
 COPY target/release/eatup_setup_server .
 ENTRYPOINT ["./eatup_setup_server"]
+
+
+# TODO copy dir when installing
